Bug 1811038 - Ensure new imap mail is seen for all folders configured to be checked. r=mkmelin,BenC
If a non-inbox folder is configured to be checked for new mail and a connection
times out or is dropped, when new messages arrive, the message count is increased
but the header for the new mail is not fetched so the new mail never appears in the
summary list. This is  especially noticed when a folder is open in a tab or window
that is not accessed often. This fixes the problem and keeps checks for new mail
in non-inbox folders using less expensive imap status instead of select.

Differential Revision: https://phabricator.services.mozilla.com/D259196