Bug 1976655 - Skip unparsable local folders in search session. r=mkmelin
When a local search encounters a folder with a missing or outdated summary
file, the search pauses and triggers a reparsing of that folder. If the
reparsing succeeds, the search resumes. However, if the reparsing fails for any
reason, the search does not handle this adequately and simply hangs. This may
happen if the summary file is corrupt or if the mbox file itself is invalid,
for example. This patch attempts to handle this situation gracefully by
skipping the affected folder and logging it in the JS console.

Differential Revision: https://phabricator.services.mozilla.com/D274578