Bug 2004030 - Add sorting support for browser.mailTabs.getListedMessages(). r=tobyp
Adds the same sorting support that was previously added to `browser.messages.list()`
in Bug 1977855.

The `getListedMessages()` function uses a fast push mechanism, which returns the
first page of messages without waiting for all messages to be retrieved. The fast
puch mechanism cannot be used when a specific sort order is requested.

Differential Revision: https://phabricator.services.mozilla.com/D275241