Bug 2003491 - Provide a simple way to expand mailing lists retrieved from address fields from the compose API. r=mkmelin
The documentation does not explain that address fields in the composer use
a fake mailbox string for mailing lists, for example `ListName <ListName>`.
With this knowledge, add-on developers *could* check all mailing lists and
find the members of the `ListName` list.

A very common post-process step for compose address fields is to use
`messengerUtils.parseMailboxString()` to get the actual email from the mailbox
string. It is therefore a convenient solution to allow that function to expand
our fake mailbox strings for mailing lists, and return the parsed members
instead of `{ name: "ListName", email: "ListName" }`. This removes a lot of
unnecessary post-processing.

Differential Revision: https://phabricator.services.mozilla.com/D274737