Bug 1979580 - Allow the user to change the open file limit on Windows. r=mkmelin
Unlike Linux/macOS, Windows requires the application to set the maximum number
of open file descriptors (`_setmaxstdio`). This patch adds support for reading
the `MOZ_MAXSTDIO` environment variable on startup, allowing users to raise the
default 512 limit up to 8192. For debugging, lowering the limit is also possible
down to the minimum accepted value of 3 (though not recommended).

Differential Revision: https://phabricator.services.mozilla.com/D274934