Bug 1877646 - Exclude saved searches from being added to unified folders. r=darktrojan
When a smart folder was created, virtual subfolders were included even though
they could not be part of the underlying search session. These subfolders did
not appear in the smart folder's properties; however, they were included in the
displayed count of selected folders. Conversely, when a new saved search was
created as a subfolder of a smart folder, it only appeared as part of the
remaining folders. The same thing happened when a saved search was renamed.

To align these behaviors, I suggest not adding virtual folders to smart folders
and letting them be part of an account's remaining folders in unified folder
mode. This will also prevent the IMAP error alert described in this bug.

Differential Revision: https://phabricator.services.mozilla.com/D248709