Bug 1953072 - Don't remove image height restrictions for zoomable images. r=aleca
I've been contempalting the best solution for this bug for a while. This exchanges having images appear too large because their size was only bound by height and our styles removing that limitation for images taking up the declared height, even if they don't need it anymore. I think this is a reasonable trade-off in the absence of a better way to enforce the aspect ratio.

Differential Revision: https://phabricator.services.mozilla.com/D275462