Bug 2001547 - Ensure sane initial conditions for compose and message windows. r=mkmelin
The test fails because the window is too small and has no area for the test to click on. This
patch removes any existing state information (like size and position) for compose and message
windows from the XulStore.

This test also prevents the "know you rights" banner to appear, which is often seen in failed
tests to have focus and apparently an open context menu, which indicates the test clicked on
the bar instead of the content. Another indicator for the bar being an issue is that these
tests do *not* fail in debug tests, where the bar is prevented automatically [1].

[1]: https://searchfox.org/comm-central/rev/669c3f6761d8c835b2e34ddd1784f610fbc7975a/mail/app/profile/all-thunderbird.js#25-28

Differential Revision: https://phabricator.services.mozilla.com/D275950