Bug 1938799 - Indicate new messages when creating a folder row. r=mkmelin
Folder rows only indicated new messages by turning blue when new messages were
received or when the folder was collapsed or expanded. Newly created rows for
folders containing new messages did not. For this reason, folders that had been
read and were added to the unread view after receiving new messages missed the
indication. The same occurred for all folders when a new three-pane tab was
opened, or folder modes were switched off and on.

Differential Revision: https://phabricator.services.mozilla.com/D276098