Bug 1925248 - Wait for content page load (and use a real page) in messageDisplayScript tests. r=tobyp
We use `awaitBrowserLoaded()` in almost all other cases when loading a content
page. The screenshot [1] of a recent test fail shows the page still loading when
the error occurs. Let's wait for the load here as well (and let's use a real
content page instead of just http://mochi.test:8888/).

[1]: https://firefoxci.taskcluster-artifacts.net/YhdqE7QyTS24rtWN3y799Q/0/public/test_info/mozilla-test-fail-screenshot_7i8qoxo4.png

Differential Revision: https://phabricator.services.mozilla.com/D276229